+++ src/Routes.js
@@ -2,6 +2,7 @@ import { createStackNavigator } from 'react-navigation';
 import Login from './screens/Login';
 import Registration from './screens/Registration';
 import MainScreen from './screens/MainScreen';
+import CreateProfile from './screens/CreateProfile';
 
 const SignedOutRoutes = createStackNavigator({
     Login:{
@@ -12,6 +13,12 @@ const SignedOutRoutes = createStackNavigator({
     },
     Registration:{
         screen:Registration,
+        navigationOptions: ({ navigation }) => ({
+            header: null,        
+        }),
+    },
+    CreateProfile:{
+        screen:CreateProfile,
         navigationOptions: ({ navigation }) => ({
             header: null,        
         }),
